requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# LangChain and RAG Framework
langchain==0.1.0
langchain-community==0.0.10
langchain-aws==0.1.0
langgraph==0.0.20

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.11.6
pdfplumber==0.10.3

# Vector Databases
faiss-cpu==1.7.4
opensearch-py==2.4.0
pinecone-client==3.0.0

# Embeddings
sentence-transformers==2.2.2

# LLM and AI
openai==1.12.0
anthropic==0.18.1

# Evaluation
ragas==0.1.0
datasets==2.17.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Web Framework (for API)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Utilities
tqdm==4.66.1
requests==2.31.0
python-multipart==0.0.6

# Logging and Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1

